+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 ## [À venir dans la prochaine version]
 ### Ajouté
 - Onboarding des nouveaux administrateurs
+- Admin : Graphique des % de temps passés sur les projets
 ### Réparé
 - Ré-affiche les notifications de saisie de temps sur le dashboard